+++ flux-RSS/src/main/java/com/lecteurfluxRSS/fluxRSS/repository/ChannelRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChannelRepository extends JpaRepository<Channel, Integer> {
+public interface ChannelRepository extends JpaRepository<Channel, Long> {
 
     @Query(value = "select * from channel ORDER BY id DESC LIMIT 1", nativeQuery=true)
     public Channel getLastRss();
+
 }
